+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 0.0.2 - 2018-07-24
+### Added
+- Daemon mode for OSC Whispers
+- Systemd unit for OSC whipsers
+### Removed
+- Verbosity and quite options for OSC Whispers
+
 ## 0.0.1 - 2018-07-23
 ### Added
 - Initial CHANGELOG and officially versioned this release.